server refactoring - 폴더 구조 - 초기화 함수들 정리
