--- src/common/error.go
@@ -1,6 +1,0 @@
-package common
-
-type ResError struct {
-	ErrType string `json:"errType,omitempty"`
-	Msg     string `json:"msg,omitempty"`
-}
