+++ src/common/error/error.go
@@ -1,0 +1,6 @@
+package error
+
+type ResError struct {
+	ErrType string `json:"errType,omitempty"`
+	Msg     string `json:"msg,omitempty"`
+}
