+++ src/docs/swagger.json
@@ -24,13 +24,13 @@
                     "400": {
                         "description": "Bad Request",
                         "schema": {
-                            "$ref": "#/definitions/common.ResError"
+                            "$ref": "#/definitions/error.ResError"
                         }
                     },
                     "500": {
                         "description": "Internal Server Error",
                         "schema": {
-                            "$ref": "#/definitions/common.ResError"
+                            "$ref": "#/definitions/error.ResError"
                         }
                     }
                 }
@@ -56,13 +56,13 @@
                     "400": {
                         "description": "Bad Request",
                         "schema": {
-                            "$ref": "#/definitions/common.ResError"
+                            "$ref": "#/definitions/error.ResError"
                         }
                     },
                     "500": {
                         "description": "Internal Server Error",
                         "schema": {
-                            "$ref": "#/definitions/common.ResError"
+                            "$ref": "#/definitions/error.ResError"
                         }
                     }
                 }
@@ -70,7 +70,7 @@
         }
     },
     "definitions": {
-        "common.ResError": {
+        "error.ResError": {
             "type": "object",
             "properties": {
                 "errType": {
