+++ src/main.go
@@ -7,7 +7,7 @@ import (
 	"github.com/labstack/echo/v4/middleware"
 	echoSwagger "github.com/swaggo/echo-swagger"
 	"main/common"
-	"main/common/db"
+	"main/common/env"
 	_ "main/docs"
 	swaggerDocs "main/docs"
 	"main/features"
@@ -15,26 +15,11 @@ import (
 )
 
 func main() {
-	if err := common.InitEnv(); err != nil {
-		fmt.Sprintf("서버 에러 발생 : %s", err.Error())
+	if err := common.InitServer(); err != nil {
+		fmt.Println(err)
 		return
 	}
-	if err := common.InitAws(); err != nil {
-		fmt.Sprintf("aws 초기화 에러 : %s", err.Error())
-		return
-	}
-	if err := common.GoogleOauthInit(); err != nil {
-		fmt.Sprintf("구글 초기화 에러 : %s", err.Error())
-		return
-	}
-	if err := db.InitMongoDB(); err != nil {
-		fmt.Sprintf("mongoDB 초기화 에러 : %s", err.Error())
-		return
-	}
-	if err := db.InitMySQL(); err != nil {
-		fmt.Sprintf("mysql 초기화 에러 : %s", err.Error())
-		return
-	}
+
 	e := echo.New()
 	e.Use(middleware.Recover())
 	e.Use(middleware.CORSWithConfig(middleware.CORSConfig{
@@ -48,13 +33,13 @@ func main() {
 	//핸드러 초기화
 	features.InitHandler(e)
 	// swagger 초기화
-	if common.Env.IsLocal {
+	if env.Env.IsLocal {
 		swaggerDocs.SwaggerInfo.Host = "localhost:3000"
 		e.GET("/swagger/*", echoSwagger.WrapHandler)
 	} else {
-		swaggerDocs.SwaggerInfo.Host = fmt.Sprintf("%s-%s.breathings.net", common.Env.Env, "ticketing")
+		swaggerDocs.SwaggerInfo.Host = fmt.Sprintf("%s-%s.breathings.net", env.Env.Env, "ticketing")
 		e.GET("/swagger/*", echoSwagger.WrapHandler)
 	}
 	e.HideBanner = true
-	e.Logger.Fatal(e.Start(fmt.Sprintf(":%s", common.Env.Port)))
+	e.Logger.Fatal(e.Start(fmt.Sprintf(":%s", env.Env.Port)))
 }
